requirements.txt
requests
Pillow
scipy
sklearn
numpy
